requirements.txt
django~=3.2
django-test-without-migrations~=0.6
dj-database-url~=0.5
python-decouple~=3.6
pillow
django-taggit~=3.0